+++ Changelog.md
@@ -3,6 +3,10 @@ List of changes of this library during time.
 
 You can get different version [here](https://www.npmjs.com/package/ngx-persian) or download them from npm command line.
 
+## V2.0.0
+1. NationalCodeService for validating Iranian national code added.
+2.NationalCodePipe(`nationalCode`) is no longer throws an Error when input is not a valid string or is not in the valid range.
+
 ## V1.1.2
 1. Adding [`getGeorgianDate` method](https://alihoseiny.github.io/ngx-persian/classes/JDate.html#getGeorgianDate) to the [JDate class](https://alihoseiny.github.io/ngx-persian/classes/JDate.html) for getting corresponding Georgian date from Jalali date object. 
 2. Fix some mistakes in the README.md file.